refactor(tools): split bloom-reference path-tracing core into tracer.rs
main.rs (2330) was one of the last two files on the 2000-line
grandfather baseline; the ray/BVH/camera/RNG/BRDF/environment/lights/
integrator core moves to tracer.rs (1252) leaving main.rs at 1088
(spec/scene/glTF loading, tone mapping, render loop, CLI). Tool builds
clean; baseline now holds only renderer/mod.rs.
