+++ tools/file-lines-baseline.json
@@ -1,4 +1,3 @@
 {
-  "native/shared/src/renderer/mod.rs": 13602,
-  "tools/bloom-reference/src/main.rs": 2330
+  "native/shared/src/renderer/mod.rs": 13602
 }
